+++ CHANGELOG.md
@@ -18,6 +18,8 @@
 - #770 Fix interops operators output types
 - #798 Fix when the algebra plan was provided using one-line as logical plan
 - #799 Fix uri values computacion in runQueryCaller
+- #792 Remove orc temp files when cached on Disk
+
 
 # BlazingSQL 0.14.0 (June 9, 2020)